@@ -1,26 +1,27 @@
-namespace CommonNet.Extensions.Tests
+﻿using FluentAssertions;
+using FsCheck.Xunit;
+using Xunit;
+
+namespace CommonNet.Extensions.Tests;
+
+public class EnumerableExtensions
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using Xunit;
-
-    public class EnumerableExtensions
+    [Fact]
+    public void Enumerable_ForEachBasicTest()
     {
-        [Fact]
-        public void Enumerable_ForEachBasicTest()
-        {
-            var data = new object[] { };
-            Assert.Throws<ArgumentNullException>(() => ((object[])null).ForEach(o => { }));
-            Assert.Throws<ArgumentNullException>(() => data.ForEach(null));
-        }
+        var data = Array.Empty<object>();
 
-        [FsCheck.Xunit.Property(MaxTest = 100, DisplayName = nameof(Enumerable_PropertyForEachTest), QuietOnSuccess = true)]
-        public void Enumerable_PropertyForEachTest(int[] data)
-        {
-            var sum = 0;
-            data.ForEach(v => sum += v);
-            Assert.Equal(sum, data.Sum());
-        }
+        Action action = () => ((object[])null!).ForEach(o => { });
+        action.Should().ThrowExactly<ArgumentNullException>();
+        action = () => data.ForEach(null!);
+        action.Should().ThrowExactly<ArgumentNullException>();
+    }
+
+    [Property(MaxTest = 100, DisplayName = nameof(Enumerable_PropertyForEachTest), QuietOnSuccess = true)]
+    public void Enumerable_PropertyForEachTest(int[] data)
+    {
+        var sum = 0;
+        data.ForEach(v => sum += v);
+        sum.Should().Be(data.Sum());
     }
 }
