@@ -1,29 +1,30 @@
-namespace CommonNet.Extensions.Tests
+﻿using System.Text;
+using FluentAssertions;
+using Xunit;
+
+namespace CommonNet.Extensions.Tests;
+
+public class StringBuilderExtensions
 {
-    using System;
-    using System.Text;
-    using Xunit;
-
-    public class StringBuilderExtensions
+    [Fact]
+    public void StringBuilde_AppendIf()
     {
-        [Fact]
-        public void StringBuilde_AppendIf()
-        {
-            const StringBuilder nullSb = null;
-            Assert.Throws<ArgumentNullException>(() => nullSb.AppendIf(1 > 0, "null"));
+        const StringBuilder? nullSb = null;
 
-            var sb = new StringBuilder();
-            var type = typeof(StringBuilderExtensions);
+        Action action = () => nullSb!.AppendIf(1 > 0, "null");
+        action.Should().ThrowExactly<ArgumentNullException>();
 
-            sb.AppendIf(type.IsNotPublic, "private")
-                .AppendIf(type.IsPublic, "public")
-                .Append(' ')
-                .AppendIf(type.IsClass, "class")
-                .AppendIf(type.IsByRef, "struct")
-                .Append(' ')
-                .Append(type.Name);
+        var sb = new StringBuilder();
+        var type = typeof(StringBuilderExtensions);
 
-            Assert.Equal($"public class {nameof(StringBuilderExtensions)}", sb.ToString());
-        }
+        sb.AppendIf(type.IsNotPublic, "internal")
+            .AppendIf(type.IsPublic, "public")
+            .Append(' ')
+            .AppendIf(type.IsClass, "class")
+            .AppendIf(type.IsByRef, "struct")
+            .Append(' ')
+            .Append(type.Name);
+
+        Assert.Equal($"public class {nameof(StringBuilderExtensions)}", sb.ToString());
     }
 }
